Revert "setup CI"
