--- .github/pull-request.yml
@@ -1,56 +1,0 @@
-name: Pull Request
-
-on: 
-  push:
-    branches:
-      - main
-      - chore-setup-CI
-  pull_request:
-    branches: 
-      - main
-      - chore-setup-CI
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-
-    strategy:
-      matrix:
-        node-version: [16.x]
-
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v2
-
-      - name: Set up Node.js ${{ matrix.node-version }}
-        uses: actions/setup-node@v2
-        with:
-          node-version: ${{ matrix.node-version }}
-          cache: 'yarn'
-
-      - name: Install dependencies
-        run: yarn install
-
-      - name: Lint
-        run: |
-          yarn lint:check
-
-      - name: Login to DockerHub
-        uses: docker/login-action@v1
-        with:
-        username: ${{ secrets.DOCKER_HUB_USERNAME }}
-        password: ${{ secrets.DOCKER_HUB_PASSWORD }} 
-      - name: Build and push
-        uses: docker/build-push-action@v2
-        with:
-          context: ./
-          file: ./Dockerfile
-          builder: ${{ steps.buildx.outputs.name }}
-          push: true
-          tags: ${{ secrets.DOCKER_HUB_USERNAME }}/autochek:latest
-          cache-from: type=registry,ref=${{ secrets.DOCKER_HUB_USERNAME }}/autochek:buildcache
-          cache-to: type=registry,ref=${{ secrets.DOCKER_HUB_USERNAME }}/autchek:buildcache,mode=max
-
-              
-
-
