--- cypress/fixtures/auth.json
@@ -1,13 +1,0 @@
-{
-  "identity": [
- {
-   "firstName":"beth",
-    "lastName":"wambui",
-    "email":"testuser",
-    "phoneNumber":"123344"
-
-
- }
-
-  ]
-}
