+++ cypress/integration/auth/verify_create_account_details.js
@@ -1,24 +1,9 @@
 /// <reference types="cypress" />
 
-import { And, Given, Then, When  } from "cypress-cucumber-preprocessor/steps";
+import { Given } from "cypress-cucumber-preprocessor/steps";
 
 Given("I visit autocheck url", function() {
   cy.visit(
-    
+    "https://marketplace.staging.myautochek.com/en/ke/auth/register"
   ).url();
 });
-
-// When("When I click on navitem to change to Ng", function(){
-    
-// });
-
-// And("I click on the register navitem", function(){
-    
-// })
-// And ("I insert information to register", function(){
-    
-// })
-// // eslint-disable-next-line cucumber/async-then
-// Then("I should register to the application", function(){
-    
-// })
