+++ cypress/integration/pageObjects/auth.js
@@ -1,0 +1,11 @@
+class AuthPage {
+  registerNavItem = "register";
+  SigninNavItem = "Sign In";
+  firstNameInputField = ".social-buttons > :nth-child(1) > div > .input";
+  LastNameInputField = ".social-buttons > :nth-child(2) > div > .input";
+  PhoneNumberInputField = ".form-control";
+  emailAddressInputField = ".form-content > :nth-child(2) > div > .input";
+  passwordInputField = ".relative > .input-container > div > .input";
+  registerButton = ".button";
+  loginInButton = ".button";
+}
